Implement if control(qubits == int): ... syntax.
The "if" statement is translated into the `control` context manager. The
context manager will consider the QComparison expression coming from the
comparison and use it to tweak the current circuit so that control lines
epected to be 0 are inverted before controlling the operations.

The context manager will restore the proper control lines after the
controlled block is executed.