Add ancilla() context for using ancilla qubits.
This allow extending the current allocation with a pool of reusable
qubits (reusability is something to be added, see TODOs in the
AncillaExtension class).

Current allocation is now tracked in a stack-like structure (which
suspiciously resembles to a stack frame) in the boilerplate package.

Future changes could include simplifying execution by removing static
analysis and assuming an execution of the current allocation (top of the
stack).